task: Move functions from profile.ts to other files
